Tx signing + custom extension of Executor type + head not registering ORA-953 ORA-1048 (#97)
Creating AlloraExecutor type extending Executor via composition.

Using the request to get env vars.
Uses new env var `ALLORA_NONCE` , that when it is in the request, that
NONCE will be signed by the node address. Flex on the validator which
may request a signature by sending an `ALLORA_NONCE` in the request.

Also removes some chain-related functionality from the heads (incl
registration as a reputer), moving it to the workers.